Integrated fully functional letterboxd scraper into scraper.py
